Fixing bug in Scoreboard related to unsigned operations
In RTL, values in the ALU are unsigned, so we need to be careful with signed
operations, such as SLT (Set less than signed) and SRA (Shift right
arithmetic).

In Testbench we are using ints, which are signed, so we need to be
careful with unsigned operations, such as SLTU (Set less than unsigned).

All other operations (ADD, SUB, XOR, OR, AND) don't care about the sign.